[MOD] : Updating the PEAR library and the associate library
The version of the libraries are :
* update ADODB 495a ;
* update PEAR 1.6.1 ;
* update PEAR module Date 1.4.7 ;
* update PEAR module DB 1.7.12 ;
* update PEAR module Auth 1.5.4 ;
* update PEAR module HTTP-Request 1.4.1 ;
* update PEAR module Net::DIME 0.3 ;
* update PEAR module Net::DNS 1.0.0 ;
* update PEAR module Net::Socket 1.0.8 ;
* update PEAR module Net::URL 1.0.15 ;
* update PEAR module XML-RPC 1.5.1 ;

I have updated the line with $Id:$ by removing the $ characters so it is now
easy to compare the official versions of the libraries to the ones in tiki.

For the Date.php I re did the modifications but not as patch but by modifying
tikidate.php (separated patch) it is cleaner and easier for futur updates.


git-svn-id: https://svn.code.sf.net/p/tikiwiki/code/trunk@8896 b456876b-0849-0410-b77d-98878d47e9d5