[FEATURE] Adding the long (too long may be) waited feature TikiTest.
This Feature allows to record a set of url and to replay it to comprae the
results. You can use Xpath expression to select the part of the page you want
to compare.

This is the first try, there might be bug hiding in the code ;p

This feature requires the tidy extension of php to validate de HTML, I plan to
use HTMLPurifier as an option but it will probabily be slower...

For the Xpath syntax look a the pages on www.w3c.org :
 http://www.w3.org/TR/xpath

As an exemple :
//div[@id='tiki-center'] select all the div from the root node where the attribute 'id' is equal to 'tiki-center'

There is a preliminary doc at :
 http://doc.tikiwiki.org/tiki-index.php?page=TikiTests

Nice testtin every body ! ;p



git-svn-id: https://svn.code.sf.net/p/tikiwiki/code/branches/1.10@12971 b456876b-0849-0410-b77d-98878d47e9d5