[NEW] comments: new 'feature_comments_moderation' to allow users with 'tiki_p_admin_comments' perm to approve (or reject) comments.
If this feature is enabled, comments are queued for the moderator and are not published until it's approbation. The moderator can approve or reject comments either directly on the comments of the object (wiki page, blog post, ...), either on the 'Admin Comments' page.

Users without 'tiki_p_admin_comments' perm won't see rejected and queued comments. They will also have the following comment: 'Your comment will have to be approved by the moderator before it's publication.'

Rewrite tiki-list_comments.php to:
	* handle table column headers (e.g. if there is only one type, 'ObjectType' is not shown anymore and 'Object' is renamed to the real type name),
	* use the global $sections* arrays,
	* add a new 'types_section' param to limit the interface to only one section (i.e. blogs, faqs, ...)

Also partially rewrite tiki-list_comments.tpl to:
	* use find.tpl,
	* better use smarty self_link,
	* add a new 'Approval' column, to approve/reject comments or see their final status (approved / rejected),
	* add a new 'Actions' column, to display/edit/delete each comments,
	* rename 'Data' column into 'Comment'

Note that it's not possible to reply to (or vote for) a comment that has not been approved, even by the moderator.

+ Add a new param 'types_tag' to find.tpl, to allow a multiple selection of types, using checkboxes (types_tag='checkbox')

+ Some CSS addons to:
	* display a "double red" left border for rejected comments,
	* display a "dotted blue" left border for queued comments (those waiting for moderation),
	* add some opacity for both rejected and queued comments,
	* fix some themes to avoid scrollbars inside comments headers/titles

+ Delete unused smarty_tiki 'ucfirst' modifier and list it in the list of php allowed functions for smarty security

+ Propagate smarty 'query', 'remarksbox', ...



git-svn-id: https://svn.code.sf.net/p/tikiwiki/code/trunk@14991 b456876b-0849-0410-b77d-98878d47e9d5