Modified logic for selection of the best language for wiki pages.
Previously, the 'best language' logic was only used if no specific wiki page 
was specified, i.e. only on the default wiki page. This meant that if you selected
a specific language version of a given page, and then followed a link to another
wiki page, you would go back to the original language, even if there was a 
translated version of the new page.

In the new version, the 'best language' is always used, unless a specific 
version has been selected via the dropdown list. Furthermore, when the user
selects a specific language, the selected language is stored in a session 
variable as his or her preferred language, and later used in the 'best language'
selection.

The update consists of the changes:
1. A hidden input field userLanguageSelected is added to translated-lang.tpl
2. The logic to set $use_best_language in tiki-index.php is moved down a couple 
   of lines, so that is applies when a specific wiki page is selected. However,
   is does not set $use_best_language if the hidden input userLanguageSelected
   is detected.
3. When the hidden input userLanguageSelected is detected, the language of 
   the selected page is stored in the session variable prefLang (in tiki-index.php).
4. In the function preferedLangs() in multilinguallib.php, if there is a 
   session variable prefLang, then the language stored in this variable is placed
   at the head of the list of preferred languages.


git-svn-id: https://svn.code.sf.net/p/tikiwiki/code/trunk@16611 b456876b-0849-0410-b77d-98878d47e9d5