[ENH] Added a controller function to be able to change a user's avatar in a bootstrap modal rather than having to navigate away and go to tiki-pick_avatar.
Needed to make some adjustments to form submission in tiki-jquery.js since the old way of serializing form data didn't work for files. The FormData call used for files is buggy in IE though so I made a check to only submit using FormData() if there was a file submitted directly in the form. 


git-svn-id: https://svn.code.sf.net/p/tikiwiki/code/trunk@58185 b456876b-0849-0410-b77d-98878d47e9d5